requirements.txt
ipdb==0.12.3
ipython==7.10.2
numpy==1.17.4
webcolors==1.12
pyautogui==0.9.53
openvino-dev==2022.1.0
